Moved to namespace symlinks
