+++ conan/tools/files/__init__.py
@@ -1,6 +1,6 @@
 from conan.tools.files.files import load, save, mkdir, ftp_download, download, get, rename, \
-    load_toolchain_args, save_toolchain_args, chdir, absolute_to_relative_symlinks, \
-    remove_external_symlinks, remove_broken_symlinks, get_symlinks
+    load_toolchain_args, save_toolchain_args, chdir
 from conan.tools.files.patches import patch, apply_conandata_patches
 from conan.tools.files.cpp_package import CppPackage
 from conan.tools.files.packager import AutoPackager
+from conan.tools.files.symlinks import symlinks
