+++ conan/tools/files/symlinks/__init__.py
@@ -1,0 +1,2 @@
+from conan.tools.files.symlinks.symlinks import absolute_to_relative_symlinks, \
+    remove_external_symlinks, remove_broken_symlinks, get_symlinks
